// src/angular/src/app/interfaces/presence.ts
export interface Presence{
    nolot: number;
    nofiche: string;
    id: string;
    dateraw: string;  
    datesem: string;  
    hdebut: string;
    hfin: string;
    codesalle: string;
    codeue: string;
    codepole: string;
    jour: string;
    enseignant: string;
    appariteur: string;
    rem_ens: string;
    rem_app: string;
    nbpresents: number;
    chaire: string;
    valide: string;
    typecours: string;
    groupe: string;
    annule: string;
    hdebut_reel: string;
    hfin_reel: string;
    datefichier: string;
    video: string;
}